refactor: update storybook action bar / alert / avatar
