+++ src/components/ActionBar/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import ActionBar from '..'
+
+export const Template: ComponentStory<typeof ActionBar> = args => (
+  <ActionBar {...args}>I am the Playground Action Bar</ActionBar>
+)
