+++ src/components/Alert/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Alert from '..'
+
+export const Template: ComponentStory<typeof Alert> = args => (
+  <Alert {...args} />
+)
