+++ src/components/Avatar/__stories__/Playground.stories.tsx
@@ -1,0 +1,8 @@
+import { Template } from './Template.stories'
+import avatar from './avatar.svg'
+
+export const Playground = Template.bind({})
+
+Playground.args = {
+  image: avatar,
+}
