+++ src/components/Avatar/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Avatar from '..'
+
+export const Template: ComponentStory<typeof Avatar> = args => (
+  <Avatar {...args} />
+)
