+++ src/components/Avatar/__stories__/TextColor.stories.tsx
@@ -1,0 +1,14 @@
+import { Template } from './Template.stories'
+
+export const TextColor = Template.bind({})
+
+TextColor.args = {
+  text: 'Hello You',
+  textColor: 'black',
+}
+
+TextColor.parameters = {
+  docs: {
+    storyDescription: 'You can set the text color by setting `textColor` prop',
+  },
+}
